pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-affiliate-system"
dynamic = ["version"]
description = "A comprehensive Django affiliate marketing and referral tracking system"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Ayodeji Akenroye", email = "aayodeji.f@gmail.com"}
]
keywords = ["django", "affiliate", "referral", "tracking", "commission", "marketing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Framework :: Django :: 5.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Office/Business :: Financial",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "Django>=4.0,<5.3",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.2.0",
]

[project.optional-dependencies]
celery = [
    "celery>=5.2.0",
    "redis>=4.5.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "django-stubs>=4.0.0",
    "djangorestframework-stubs>=3.14.0",
    "tox>=4.0.0",
    "twine>=4.0.0",
    "wheel>=0.40.0",
    "build>=0.10.0",
    "coverage>=7.0.0",
    "factory-boy>=3.2.0",
    "faker>=18.0.0",
]
all = [
    "celery>=5.2.0",
    "redis>=4.5.0",
]

[project.urls]
Homepage = "https://github.com/aayodejii/django-affiliate-system"
Documentation = "https://github.com/aayodejii/django-affiliate-system#readme"
Repository = "https://github.com/aayodejii/django-affiliate-system"
"Bug Reports" = "https://github.com/aayodejii/django-affiliate-system/issues"

[tool.setuptools.packages.find]
exclude = ["tests*", "docs*"]

[tool.setuptools]
include-package-data = true
zip-safe = false

[tool.setuptools.dynamic]
version = {attr = "django_affiliate_system.__version__"}

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "example_project.example_project.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100
